AI-Smart-Mirror/requirements.txt
speechrecognition
gtts
py4j
python-dateutil
pydub
feedparser
numpy
arabic-reshaper
requests
python-bidi
opencv-python